+++ src/index.js
@@ -5,6 +5,7 @@ export * from './components/Checkbox';
 export * from './components/CodeBlock';
 export * from './components/ConfirmModal';
 export * from './components/ControlsPanel';
+export * from './components/CopyToClipboard';
 export * from './components/Divider'; // Is it necessary component?
 export * from './components/DotIndicator';
 export * from './components/FlatList';