+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added withTooltip HOC
 
+- Added CopyToClipboard component and copyToClipboard function
+
 ## [0.12.0] - 2020-02-10
 
 - Updated Markdown styles